+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 17. Fix crash in anonymous layout that starts with whitespace (occurred e.g. on Wikipedia)
 18. Screen scaling (DPI) solves [the text issue](https://github.com/usadson/retina/issues/23)
 19. CSS [`cursor`](https://drafts.csswg.org/css-ui/#cursor) property is now supported
+20. Relative `font-size`s (e.g. the `em` unit) are not resolved at compute time, so that they cascade correctly.
 
 ### Performance improvements
 1. `ColorMaterialRenderer` and `TextureMaterialRenderer` are now globally shared instead of per `Artwork`, which previously made the creation time of tiles significantly slower.